Making progress on geneticAlgorithm. Reproduce function is almost finished. Changed numIndividuals and numGenes in data.h to long from int

+++ data.h
@@ -3,8 +3,8 @@
 
 struct data {
 
-	int numIndividuals;
-	int numGenes;
+	long numIndividuals;
+	long numGenes;
 	double ** values;
 
 };
